+++ modules/planning/open_space/node3d.cc
@@ -40,9 +40,11 @@ Node3d::Node3d(double x, double y, double phi,
   phi_grid_ = static_cast<std::size_t>(
       (phi_ - (-M_PI)) /
       open_space_conf.warm_start_config().phi_grid_resolution());
-  index_ =
-      phi_grid_ * (XYbounds[1] - XYbounds[0]) * (XYbounds[3] - XYbounds[2]) +
-      y_grid_ * (XYbounds[1] - XYbounds[0]) + x_grid_;
+  index_ = static_cast<std::size_t>
+      (static_cast<double>(phi_grid_) * (XYbounds[1] - XYbounds[0]) *
+      (XYbounds[3] - XYbounds[2]) +
+      static_cast<double>(y_grid_) * (XYbounds[1] - XYbounds[0]) +
+      static_cast<double>(x_grid_));
 }
 
 Node3d::Node3d(double x, double y, double phi) {
@@ -71,9 +73,11 @@ Node3d::Node3d(double x, double y, double phi, std::vector<double> traversed_x,
   phi_grid_ = static_cast<std::size_t>(
       (phi_ - (-M_PI)) /
       open_space_conf.warm_start_config().phi_grid_resolution());
-  index_ =
-      phi_grid_ * (XYbounds[1] - XYbounds[0]) * (XYbounds[3] - XYbounds[2]) +
-      y_grid_ * (XYbounds[1] - XYbounds[0]) + x_grid_;
+  index_ = static_cast<std::size_t>
+      (static_cast<double>(phi_grid_) * (XYbounds[1] - XYbounds[0]) *
+      (XYbounds[3] - XYbounds[2]) +
+      static_cast<double>(y_grid_) * (XYbounds[1] - XYbounds[0]) +
+      static_cast<double>(x_grid_));
   traversed_x_ = traversed_x;
   traversed_y_ = traversed_y;
   traversed_phi_ = traversed_phi;
